proveale/OLD/requirements.txt
accelerate==1.7.0aiohappyeyeballs==2.6.1aiohttp==3.12.13aioice==0.10.1aiortc==1.13.0aiosignal==1.3.2altair==5.5.0annotated-types==0.7.0anyio==4.9.0attrs==25.3.0av==14.4.0beautifulsoup4==4.13.4bitsandbytes==0.46.0blinker==1.9.0branca==0.8.1bs4==0.0.2cachetools==6.1.0certifi==2025.6.15cffi==1.17.1chardet==5.2.0charset-normalizer==3.4.2click==8.2.1colorama==0.4.6cryptography==45.0.3dnspython==2.7.0eval_type_backport==0.2.2Faker==37.4.0ffmpeg-python==0.2.0filelock==3.18.0Flask==3.1.1folium==0.20.0frozenlist==1.7.0fsspec==2025.5.1future==1.0.0gitdb==4.0.12GitPython==3.1.44google-crc32c==1.7.1greenlet==3.2.3h11==0.16.0hf-xet==1.1.4httpcore==1.0.9httpx==0.28.1huggingface-hub==0.33.0ibm-cos-sdk==2.14.1ibm-cos-sdk-core==2.14.1ibm-cos-sdk-s3transfer==2.14.1ibm_watsonx_ai==1.3.20idna==3.10ifaddr==0.2.0itsdangerous==2.2.0Jinja2==3.1.6jmespath==1.0.1jsonpatch==1.33jsonpointer==3.0.0jsonschema==4.24.0jsonschema-specifications==2025.4.1langchain-core==0.3.61langchain-ibm==0.3.11langsmith==0.3.42llvmlite==0.44.0lomond==0.3.3markdown-it-py==3.0.0MarkupSafe==3.0.2mdurl==0.1.2more-itertools==10.7.0mpmath==1.3.0multidict==6.5.0narwhals==1.42.1networkx==3.5numba==0.61.2numpy==2.2.6openai-whisper==20240930orjson==3.10.18packaging==25.0pandas==2.3.0pdfrw==0.4pillow==11.2.1playwright==1.52.0prettytable==3.16.0propcache==0.3.2protobuf==6.31.1psutil==7.0.0pyarrow==20.0.0pycparser==2.22pydantic==2.11.5pydantic_core==2.33.2pydeck==0.9.1pyecharts==2.0.8pyee==13.0.0Pygments==2.19.1pylibsrtp==0.12.0pymongo==4.13.2pyOpenSSL==25.1.0pypdf==5.6.0python-codicefiscale==0.10.3python-dateutil==2.9.0.post0python-fsutil==0.15.0python-slugify==8.0.4pytz==2025.2PyYAML==6.0.2referencing==0.36.2regex==2024.11.6reportlab==4.4.2requests==2.32.4requests-toolbelt==1.0.0rich==14.0.0rpds-py==0.25.1safetensors==0.5.3setuptools==80.9.0shellingham==1.5.4simplejson==3.20.1six==1.17.0smmap==5.0.2sniffio==1.3.1soupsieve==2.7streamlit==1.46.0streamlit-echarts==0.4.0streamlit_folium==0.25.0sympy==1.14.0tabulate==0.9.0tenacity==9.1.2text-unidecode==1.3tiktoken==0.9.0together==1.5.16tokenizers==0.21.1toml==0.10.2torch==2.7.1tornado==6.5.1tqdm==4.67.1transformers==4.52.4typer==0.15.4typing-inspection==0.4.1typing_extensions==4.14.0tzdata==2025.2urllib3==2.5.0watchdog==6.0.0wcwidth==0.2.13Werkzeug==3.1.3xyzservices==2025.4.0yarl==1.20.1zstandard==0.23.0